refactor counter connect and prop maps to container
