+++ src/components/Counter.js
@@ -1,27 +1,4 @@
 import React, { Component } from 'react';
-import {connect} from 'react-redux';
-import store from '../redux/store';
-
-const mapStateToProps = () => {
-  return {
-    value: store.getState()
-  };
-};
-
-const mapDispatchToProps = () => {
-  return {
-    dispatchIncrement () {
-      store.dispatch({
-        type: 'INCREMENT'
-      })
-    },
-    dispatchDecrement () {
-      store.dispatch({
-        type: 'DECREMENT'
-      })
-    }
-  }
-};
 
 class Counter extends Component {
   render() {
@@ -35,4 +12,4 @@ class Counter extends Component {
   }
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(Counter);
+export default Counter;
